Search banner basics

+++ app/src/main/java/com/melodie/parotia/model/Photo.kt
@@ -35,8 +35,8 @@ data class Location(
 )
 
 data class Position(
-    val latitude: Long,
-    val longitude: Long
+    val latitude: Double,
+    val longitude: Double
 )
 
 data class Tag(
